[Java] Migrate `RecordingDescriptorHeader.valid` field from byte to enum and rename it to `state`.
