+++ aeron-archive/src/main/java/io/aeron/archive/ArchiveMigration_1_2.java
@@ -20,12 +20,12 @@ import io.aeron.archive.codecs.RecordingDescriptorEncoder;
 import io.aeron.archive.codecs.RecordingDescriptorHeaderDecoder;
 import io.aeron.archive.codecs.RecordingDescriptorHeaderEncoder;
 import org.agrona.AsciiEncoding;
-import org.agrona.CloseHelper;
 import org.agrona.LangUtil;
 import org.agrona.SemanticVersion;
 import org.agrona.collections.ArrayUtil;
 
 import java.io.File;
+import java.io.IOException;
 import java.io.PrintStream;
 import java.nio.channels.FileChannel;
 import java.nio.file.Files;
@@ -33,8 +33,8 @@ import java.nio.file.Path;
 import java.nio.file.attribute.FileTime;
 import java.util.Arrays;
 
-import static io.aeron.archive.Catalog.INVALID;
 import static io.aeron.archive.MigrationUtils.fullVersionString;
+import static io.aeron.archive.codecs.RecordingState.INVALID;
 
 class ArchiveMigration_1_2 implements ArchiveMigrationStep
 {
@@ -51,38 +51,41 @@ class ArchiveMigration_1_2 implements ArchiveMigrationStep
         final Catalog catalog,
         final File archiveDir)
     {
-        final FileChannel migrationTimestampFile = MigrationUtils.createMigrationTimestampFile(
-            archiveDir, markFile.decoder().version(), minimumVersion());
-
-        catalog.forEach(
-            (headerEncoder, headerDecoder, encoder, decoder) ->
-            {
-                final String version1Prefix = decoder.recordingId() + "-";
-                final String version1Suffix = ".rec";
-                String[] segmentFiles = archiveDir.list(
-                    (dir, filename) -> filename.startsWith(version1Prefix) && filename.endsWith(version1Suffix));
-
-                if (null == segmentFiles)
+        try (FileChannel ignore = MigrationUtils.createMigrationTimestampFile(
+            archiveDir, markFile.decoder().version(), minimumVersion()))
+        {
+            catalog.forEach(
+                (headerEncoder, headerDecoder, encoder, decoder) ->
                 {
-                    segmentFiles = ArrayUtil.EMPTY_STRING_ARRAY;
-                }
+                    final String version1Prefix = decoder.recordingId() + "-";
+                    final String version1Suffix = ".rec";
+                    String[] segmentFiles = archiveDir.list(
+                        (dir, filename) -> filename.startsWith(version1Prefix) && filename.endsWith(version1Suffix));
 
-                migrateRecording(
-                    stream,
-                    archiveDir,
-                    segmentFiles,
-                    version1Prefix,
-                    version1Suffix,
-                    headerEncoder,
-                    headerDecoder,
-                    encoder,
-                    decoder);
-            });
+                    if (null == segmentFiles)
+                    {
+                        segmentFiles = ArrayUtil.EMPTY_STRING_ARRAY;
+                    }
 
-        markFile.encoder().version(minimumVersion());
-        catalog.updateVersion(minimumVersion());
+                    migrateRecording(
+                        stream,
+                        archiveDir,
+                        segmentFiles,
+                        version1Prefix,
+                        version1Suffix,
+                        headerEncoder,
+                        headerDecoder,
+                        encoder,
+                        decoder);
+                });
 
-        CloseHelper.close(migrationTimestampFile);
+            markFile.encoder().version(minimumVersion());
+            catalog.updateVersion(minimumVersion());
+        }
+        catch (final IOException ex)
+        {
+            LangUtil.rethrowUnchecked(ex);
+        }
     }
 
     private void migrateRecording(
@@ -102,7 +105,7 @@ class ArchiveMigration_1_2 implements ArchiveMigrationStep
         final long segmentLength = decoder.segmentFileLength();
         final long startTermBasePosition = startPosition - (startPosition & (termLength - 1));
 
-        if (headerDecoder.valid() == INVALID)
+        if (headerDecoder.state() == INVALID)
         {
             return;
         }
