+++ aeron-archive/src/main/resources/archive/aeron-archive-codecs.xml
@@ -2,7 +2,7 @@
 <sbe:messageSchema xmlns:sbe="http://fixprotocol.io/2016/sbe"
                    package="io.aeron.archive.codecs"
                    id="101"
-                   version="4"
+                   version="5"
                    semanticVersion="5.2"
                    description="Message Codecs for communicating with an Aeron Archive."
                    byteOrder="littleEndian">
@@ -47,6 +47,10 @@
             <validValue name="REPLICATE" description="Recording descriptor replicated from source archive.">3</validValue>
             <validValue name="MERGE" description="Recording merged with live stream after replay.">4</validValue>
             <validValue name="SYNC" description="Recording synchronised with source archive.">5</validValue>
+        </enum>
+        <enum name="RecordingState" encodingType="int32" description="State of a recording in the Catalog.">
+            <validValue name="INVALID" description="Recording is invalid.">0</validValue>
+            <validValue name="VALID" description="Recording is valid.">1</validValue>
         </enum>
         <type name="time_t" primitiveType="int64" description="Epoch time in milliseconds since 1 Jan 1970 UTC."/>
         <type name="version_t" primitiveType="int32" presence="optional" nullValue="0" minValue="2" maxValue="16777215"
@@ -329,7 +333,7 @@
                  id="21"
                  description="Used in the catalog to describe the recording descriptor entry which follows.">
         <field name="length"               id="1" type="int32"/>
-        <field name="valid"                id="2" type="int8"/>
+        <field name="state"                id="2" type="RecordingState" description="State of the recording."/>
         <field name="reserved"             id="3" type="int8" offset="31"/>
     </sbe:message>
